Simplify role handling for actor groups
