+++ apps/hash-api/src/graph/account-permission-management.ts
@@ -7,36 +7,63 @@ import type {
   WebId,
 } from "@blockprotocol/type-system";
 import { NotFoundError } from "@local/hash-backend-utils/error";
+import type {
+  RoleAssignmentStatus,
+  RoleUnassignmentStatus,
+} from "@local/hash-graph-client";
 
 import type { ImpureGraphFunction } from "./context-types";
 
+export const getActorGroupMembers: ImpureGraphFunction<
+  { actorGroupId: ActorGroupEntityUuid },
+  Promise<ActorEntityUuid[]>
+> = async ({ graphApi }, authentication, params) =>
+  graphApi
+    .getActorGroupRoleAssignments(
+      authentication.actorId,
+      params.actorGroupId,
+      "member",
+    )
+    .then(({ data }) => data as ActorEntityUuid[]);
+
 export const addActorGroupMember: ImpureGraphFunction<
   { actorId: ActorEntityUuid; actorGroupId: ActorGroupEntityUuid },
-  Promise<boolean>
-> = async ({ graphApi }, { actorId }, params) => {
-  await graphApi.assignActorGroupRole(
-    actorId,
-    params.actorGroupId,
-    "member",
-    params.actorId,
-  );
+  Promise<RoleAssignmentStatus>
+> = async ({ graphApi }, authentication, params) =>
+  graphApi
+    .assignActorGroupRole(
+      authentication.actorId,
+      params.actorGroupId,
+      "member",
+      params.actorId,
+    )
+    .then(({ data }) => data);
 
-  return true;
-};
+export const isActorGroupMember: ImpureGraphFunction<
+  { actorId: ActorEntityUuid; actorGroupId: ActorGroupEntityUuid },
+  Promise<boolean>
+> = async ({ graphApi }, authentication, params) =>
+  graphApi
+    .hasActorGroupRole(
+      authentication.actorId,
+      params.actorGroupId,
+      "member",
+      params.actorId,
+    )
+    .then(({ data }) => data);
 
 export const removeActorGroupMember: ImpureGraphFunction<
   { actorId: ActorEntityUuid; actorGroupId: ActorGroupEntityUuid },
-  Promise<boolean>
-> = async ({ graphApi }, { actorId }, params) => {
-  await graphApi.unassignActorGroupRole(
-    actorId,
-    params.actorGroupId,
-    "member",
-    params.actorId,
-  );
-
-  return true;
-};
+  Promise<RoleUnassignmentStatus>
+> = async ({ graphApi }, authentication, params) =>
+  graphApi
+    .unassignActorGroupRole(
+      authentication.actorId,
+      params.actorGroupId,
+      "member",
+      params.actorId,
+    )
+    .then(({ data }) => data);
 
 export const createUserActor: ImpureGraphFunction<
   {
