+++ tests/hash-backend-integration/src/tests/graph/knowledge/system-types/hash-instance.test.ts
@@ -1,15 +1,17 @@
-import { ensureSystemGraphIsInitialized } from "@apps/hash-api/src/graph/ensure-system-graph-is-initialized";
 import {
-  addHashInstanceAdmin,
-  getHashInstanceGroupMembers,
-  removeHashInstanceAdmin,
-} from "@apps/hash-api/src/graph/knowledge/system-types/hash-instance";
+  addActorGroupMember,
+  getActorGroupMembers,
+  isActorGroupMember,
+  removeActorGroupMember,
+} from "@apps/hash-api/src/graph/account-permission-management";
+import { ensureSystemGraphIsInitialized } from "@apps/hash-api/src/graph/ensure-system-graph-is-initialized";
 import type { User } from "@apps/hash-api/src/graph/knowledge/system-types/user";
 import { systemAccountId } from "@apps/hash-api/src/graph/system-account";
+import type { TeamId } from "@blockprotocol/type-system";
 import type { HashInstance } from "@local/hash-backend-utils/hash-instance";
 import {
   getHashInstance,
-  isUserHashInstanceAdmin,
+  getInstanceAdminsTeam,
 } from "@local/hash-backend-utils/hash-instance";
 import { Logger } from "@local/hash-backend-utils/logger";
 import { publicUserAccountId } from "@local/hash-backend-utils/public-user-account-id";
@@ -46,6 +48,16 @@ describe("Hash Instance", () => {
     expect(hashInstance).toBeTruthy();
   });
 
+  let instanceAdminTeamId: TeamId;
+
+  it("can get the instance admin team id", async () => {
+    const { teamId } = await getInstanceAdminsTeam(graphContext, {
+      actorId: systemAccountId,
+    });
+    instanceAdminTeamId = teamId;
+    expect(instanceAdminTeamId).toBeTruthy();
+  });
+
   let testHashInstanceAdmin: User;
   let authentication: AuthenticationContext;
 
@@ -58,78 +70,90 @@ describe("Hash Instance", () => {
     authentication = { actorId: testHashInstanceAdmin.accountId };
 
     expect(
-      await isUserHashInstanceAdmin(graphContext, authentication, {
-        userAccountId: testHashInstanceAdmin.accountId,
+      await isActorGroupMember(graphContext, authentication, {
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       }),
     ).toBeFalsy();
   });
 
   it("can add a hash instance admin", async () => {
     expect(
-      await getHashInstanceGroupMembers(
+      await getActorGroupMembers(
         graphContext,
         {
           actorId: systemAccountId,
         },
-        {},
+        {
+          actorGroupId: instanceAdminTeamId,
+        },
       ),
     ).not.toContain(testHashInstanceAdmin.accountId);
 
     expect(
-      await isUserHashInstanceAdmin(graphContext, authentication, {
-        userAccountId: testHashInstanceAdmin.accountId,
+      await isActorGroupMember(graphContext, authentication, {
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       }),
     ).toBeFalsy();
 
-    await addHashInstanceAdmin(
+    await addActorGroupMember(
       graphContext,
       { actorId: systemAccountId },
       {
-        user: testHashInstanceAdmin,
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       },
     );
   });
 
   it("can get the hash instance group members", async () => {
     expect(
-      await isUserHashInstanceAdmin(graphContext, authentication, {
-        userAccountId: testHashInstanceAdmin.accountId,
+      await isActorGroupMember(graphContext, authentication, {
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       }),
     ).toBeTruthy();
 
     expect(
-      await getHashInstanceGroupMembers(
+      await getActorGroupMembers(
         graphContext,
         {
           actorId: systemAccountId,
         },
-        {},
+        {
+          actorGroupId: instanceAdminTeamId,
+        },
       ),
     ).toContain(testHashInstanceAdmin.accountId);
   });
 
   it("can remove a hash instance admin", async () => {
-    await removeHashInstanceAdmin(
+    await removeActorGroupMember(
       graphContext,
       { actorId: systemAccountId },
       {
-        user: testHashInstanceAdmin,
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       },
     );
 
     expect(
-      await getHashInstanceGroupMembers(
+      await getActorGroupMembers(
         graphContext,
         {
           actorId: systemAccountId,
         },
-        {},
+        {
+          actorGroupId: instanceAdminTeamId,
+        },
       ),
     ).not.toContain(testHashInstanceAdmin.accountId);
 
     expect(
-      await isUserHashInstanceAdmin(graphContext, authentication, {
-        userAccountId: testHashInstanceAdmin.accountId,
+      await isActorGroupMember(graphContext, authentication, {
+        actorId: testHashInstanceAdmin.accountId,
+        actorGroupId: instanceAdminTeamId,
       }),
     ).toBeFalsy();
   });
